requirements.txt
pytest-playwright
playwright
pandas
matplotlib
st-supabase-connection
requests
openai
beautifulsoup4
streamlit-supabase-auth
streamlit_antd_components
plotly
uuid
